+++ lib/main.dart
@@ -4,9 +4,9 @@ import 'package:places/data/interactor/favorite_sights.dart';
 import 'package:places/res/app_routes.dart';
 import 'package:places/res/themes.dart';
 import 'package:places/ui/screen/map_screen.dart';
-import 'package:places/ui/screen/sight_list_screen.dart';
+import 'package:places/ui/screen/places_list_screen.dart';
 import 'package:places/ui/screen/splash_screen.dart';
-import 'package:places/ui/screen/visiting_screen.dart';
+import 'package:places/ui/screen/favorites.dart';
 import 'package:places/ui/screen/settings_screen.dart';
 import 'package:places/mocks.dart';
 import 'package:provider/provider.dart';
