+++ lib/ui/screen/place_details_screen.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-import 'package:places/data/model/sight.dart';
+import 'package:places/data/model/place.dart';
 import 'package:places/res/icons.dart';
 import 'package:places/res/text_styles.dart';
 import 'package:places/res/text_strings.dart';
